Turn everything into components
