+++ src/components/EntrantName.jsx
@@ -1,0 +1,6 @@
+const EntrantName = (props) => {
+  const { name } = props;
+  return <div className="entrant-name">{name}</div>;
+};
+
+export default EntrantName;
