+++ src/scripts/util.js
@@ -1,18 +1,7 @@
-function createLadder(num) {
-  const ladderArray = [];
-  for (let i = 0; i < num; i++) {
-    ladderArray.push(EMPTY_LADDER_ENTRY);
-  }
-  return ladderArray;
-}
-
-function formatDifference(difference) {
+export const formatDifference = (difference) => {
   return difference === 0
     ? "--"
     : difference > 0
     ? `+${difference}`
     : `${difference}`;
-}
-
-exports.createLadder = createLadder;
-exports.formatDifference = formatDifference;
+};
